Canonical Approach with latest branch
This PR changes the current version docs to latest and adds
canonical links to the old docs (where it is easily possible).

Instead of copying the folder and name it for the new version the
process would look like:
1. copying latest for the old version
2. change the alias in index.md of latest to the new version

All links will stay the same, the only issue which might arise is
the CLI update, which we would need to investigate specially.

Also this allows for easier version specific changes/updates.

Signed-off-by: Simon Schrottner <simon.schrottner@dynatrace.com>