Poor Man Fix for missing images
Hugo detects which assets are used, and which ones not. As we only
have a reference to the images within JS, Hugo does not detect
their usage and is not copying them over to the right directory
when building the page.
We by-pass this by adding them to the static folder in the right
directory.

Signed-off-by: Simon Schrottner <simon.schrottner@dynatrace.com>