Exploration with Docsy
Docsy is a widely used documentation theme, and it offers a lot of
features we are looking for, specially versioning etc.

Sadly it still does not offer multipage support, but at least we
are heading into the right direction, and it seems like an easy
fix for a lot of our problems.

Signed-off-by: Simon Schrottner <simon.schrottner@dynatrace.com>