Remove ruxit agent
The ruxit agent configuration was outdated, the host it tried to
connect to was unreachable. Althought this data might be
interesting, we also do have google analytics and scarf to also
generate this information.

If we want to enable the ruxit agent again, we need to think about
which host we are using for reporting.

Signed-off-by: Simon Schrottner <simon.schrottner@dynatrace.com>