Replacing hardcoded version with param
As we do now have a param to get the latest version, we can use
this wihin our concepts, install, and explore to link to the
most up-to-date documentation part. Instead of a hardcoded version
which needs to be changed regularly. As a result we do have a lot
of links pointing to outdated versions.

Signed-off-by: Simon Schrottner <simon.schrottner@dynatrace.com>