Sitemap clean up
Our sitemap contained too many pages, also those which are
refering to a more up to date version. This is/was problematic
for SEO. Search engines do see the sitemap as source of truth and
get confused if there are pages, which declare themselves as
outdated. Hence that, we are removing all outdated versions from
the sitemap.xml

Signed-off-by: Simon Schrottner <simon.schrottner@dynatrace.com>